requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.1
langchain-core==0.1.0
faiss-cpu==1.7.4
huggingface-hub==0.18.0
sentence-transformers==2.2.2
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
python-dotenv==1.0.0
requests==2.31.0
